pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = "Dependencies to run the project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.3"
httpx = "^0.27.2"
openai = "^1.52.0"
pydantic = "^2.9.2"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
aioredis = "^2.0.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
black = "^24.10.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
